APPLE: fix Malta version daemon installation
